PRSD-835: Moves task-related view models into taskModels
