+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/models/viewModels/taskModels/TaskListViewModel.kt
@@ -1,4 +1,4 @@
-package uk.gov.communities.prsdb.webapp.models.viewModels
+package uk.gov.communities.prsdb.webapp.models.viewModels.taskModels
 
 data class TaskListViewModel(
     val title: String,
