+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/models/viewModels/taskModels/TaskSectionViewModel.kt
@@ -1,4 +1,4 @@
-package uk.gov.communities.prsdb.webapp.models.viewModels
+package uk.gov.communities.prsdb.webapp.models.viewModels.taskModels
 
 data class TaskSectionViewModel(
     val headingKey: String,
