+++ src/main/resources/templates/fragments/taskList/taskListStatus.html
@@ -1,9 +1,10 @@
-<!--/*@thymesVar id="status" type="uk.gov.communities.prsdb.webapp.models.viewModels.TaskStatusViewModel"*/-->
+<!--/*@thymesVar id="status" type="uk.gov.communities.prsdb.webapp.models.viewModels.taskModels.TaskStatusViewModel"*/-->
 <div th:fragment="taskListStatus(status, id)"
      class="govuk-task-list__status"
      th:classappend="${status.isCannotStart} ? 'govuk-task-list__status--cannot-start-yet' : ''"
-     th:id="${id}" >
-    <strong th:remove="${status.isTag} ? none : tag" th:text="#{${status.textKey}}" th:classappend="${status.tagClass}" class="govuk-tag">
+     th:id="${id}">
+    <strong th:remove="${status.isTag} ? none : tag" th:text="#{${status.textKey}}" th:classappend="${status.tagClass}"
+            class="govuk-tag">
         status.text
     </strong>
 </div>
